# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LinguaReader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vocab:
    get:
      operationId: getVocab
      tags: [vocab]
      summary: Get all vocabulary entries for the user
      parameters:
        - in: query
          name: status
          schema:
            type: ["integer", "null"]
        - in: query
          name: limit
          schema:
            type: ["integer", "null"]
        - in: query
          name: offset
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Vocabulary list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VocabEntry"
...
  /vocab/lookup:
    post:
      operationId: lookupVocab
      tags: [vocab]
      summary: Look up vocab state for a list of dictionary forms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabLookupInput"
      responses:
        "200":
          description: Map of dictionary_form → VocabEntry
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/VocabEntry"
  /vocab/{id}:
    patch:
      operationId: updateVocabEntry
      tags: [vocab]
      summary: Update vocab state or note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabUpdate"
      responses:
        "200":
          description: Updated vocab entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabEntry"
  /vocab/upsert:
    post:
      operationId: upsertVocab
      tags: [vocab]
      summary: Create or update a vocab entry (tap on word)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabUpsertInput"
      responses:
        "200":
          description: Upserted vocab entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabEntry"

  # ─── DICTIONARY ─────────────────────────────────────────────────────────
...
  /srs/review:
    post:
      operationId: submitSrsReview
      tags: [srs]
      summary: Submit a review result (SM-2 update)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SrsReviewInput"
      responses:
        "200":
          description: Updated vocab entry with next review date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabEntry"

  # ─── STATS ──────────────────────────────────────────────────────────────
...
  /stats/history:
    get:
      operationId: getStatsHistory
      tags: [stats]
      summary: Get daily stats history for charting known-word growth
      parameters:
        - in: query
          name: days
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Daily stats array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStats"
  /stats/record:
    post:
      operationId: recordStats
      tags: [stats]
      summary: Record a reading/listening session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatsSessionInput"
      responses:
        "200":
          description: Updated daily stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyStats"

  # ─── IMPORT ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    VocabEntry:
      type: object
      required: [id, surface, dictionaryForm, reading, status, timesSeen, createdAt, updatedAt]
      properties:
        id:
          type: string
        surface:
          type: string
        dictionaryForm:
          type: string
        reading:
          type: string
        status:
          type: integer
          description: "0=unknown, 1-4=learning stages, 5=known, 99=ignored"
        timesSeen:
          type: integer
        note:
          type: ["string", "null"]
        sourceLessonId:
          type: ["string", "null"]
        nextReview:
          type: ["string", "null"]
        easeFactor:
          type: ["number", "null"]
        intervalDays:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    VocabUpdate:
      type: object
      properties:
        status:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
    VocabUpsertInput:
      type: object
      required: [surface, dictionaryForm, reading]
      properties:
        surface:
          type: string
        dictionaryForm:
          type: string
        reading:
          type: string
        status:
          type: ["integer", "null"]
        sourceLessonId:
          type: ["string", "null"]
    VocabLookupInput:
      type: object
      required: [dictionaryForms]
      properties:
        dictionaryForms:
          type: array
          items:
            type: string
...
    SrsReviewInput:
      type: object
      required: [vocabId, quality]
      properties:
        vocabId:
          type: string
        quality:
          type: integer
          description: "SM-2 quality rating 0-5"

    # ── Stats ─────────────────────────────────────────────────────────────
...
    DailyStats:
      type: object
      required: [date, wordsRead, listeningMinutes, newLingqs, knownWordsTotal, studyMinutes]
      properties:
        date:
          type: string
        wordsRead:
          type: integer
        listeningMinutes:
          type: integer
        newLingqs:
          type: integer
        knownWordsTotal:
          type: integer
        studyMinutes:
          type: integer
    StatsSessionInput:
      type: object
      properties:
        wordsRead:
          type: ["integer", "null"]
        listeningMinutes:
          type: ["integer", "null"]
        newLingqs:
          type: ["integer", "null"]
        studyMinutes:
          type: ["integer", "null"]

    # ── Import ────────────────────────────────────────────────────────────
...