# lib/api-zod/src/generated/types/vocabEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */

export interface VocabEntry {
  id: string;
  surface: string;
  dictionaryForm: string;
  reading: string;
  /** 0=unknown, 1-4=learning stages, 5=known, 99=ignored */
  status: number;
  timesSeen: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  sourceLessonId?: string | null;
  /** @nullable */
  nextReview?: string | null;
  /** @nullable */
  easeFactor?: number | null;
  /** @nullable */
  intervalDays?: number | null;
  createdAt: string;
  updatedAt: string;
}