# lib/api-zod/src/generated/types/vocabUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */

export interface VocabUpdate {
  /** @nullable */
  status?: number | null;
  /** @nullable */
  note?: string | null;
}